+++ apps/client/src/app/components/access-table/access-table.module.ts
@@ -5,6 +5,7 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatMenuModule } from '@angular/material/menu';
 import { MatTableModule } from '@angular/material/table';
 import { RouterModule } from '@angular/router';
+import { IonIcon } from '@ionic/angular/standalone';
 
 import { AccessTableComponent } from './access-table.component';
 
@@ -14,6 +15,7 @@ import { AccessTableComponent } from './access-table.component';
   imports: [
     ClipboardModule,
     CommonModule,
+    IonIcon,
     MatButtonModule,
     MatMenuModule,
     MatTableModule,