+++ apps/client/src/app/components/admin-overview/admin-overview.module.ts
@@ -11,6 +11,7 @@ import { MatSelectModule } from '@angular/material/select';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { RouterModule } from '@angular/router';
+import { IonIcon } from '@ionic/angular/standalone';
 
 import { AdminOverviewComponent } from './admin-overview.component';
 
@@ -21,6 +22,7 @@ import { AdminOverviewComponent } from './admin-overview.component';
     CommonModule,
     FormsModule,
     GfValueComponent,
+    IonIcon,
     MatButtonModule,
     MatCardModule,
     MatMenuModule,