+++ apps/client/src/app/components/home-watchlist/home-watchlist.component.ts
@@ -23,6 +23,9 @@ import {
 import { MatButtonModule } from '@angular/material/button';
 import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router, RouterModule } from '@angular/router';
+import { IonIcon } from '@ionic/angular/standalone';
+import { addIcons } from 'ionicons';
+import { addOutline } from 'ionicons/icons';
 import { DeviceDetectorService } from 'ngx-device-detector';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
@@ -36,6 +39,7 @@ import { CreateWatchlistItemDialogParams } from './create-watchlist-item-dialog/
     CommonModule,
     GfBenchmarkComponent,
     GfPremiumIndicatorComponent,
+    IonIcon,
     MatButtonModule,
     RouterModule
   ],
@@ -103,6 +107,8 @@ export class HomeWatchlistComponent implements OnDestroy, OnInit {
           this.changeDetectorRef.markForCheck();
         }
       });
+
+    addIcons({ addOutline });
   }
 
   public ngOnInit() {