+++ apps/client/src/app/pages/blog/blog-page.module.ts
@@ -1,13 +1,14 @@
 import { CommonModule } from '@angular/common';
 import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
 import { MatCardModule } from '@angular/material/card';
+import { IonIcon } from '@ionic/angular/standalone';
 
 import { BlogPageRoutingModule } from './blog-page-routing.module';
 import { BlogPageComponent } from './blog-page.component';
 
 @NgModule({
   declarations: [BlogPageComponent],
-  imports: [BlogPageRoutingModule, CommonModule, MatCardModule],
+  imports: [BlogPageRoutingModule, CommonModule, IonIcon, MatCardModule],
   schemas: [CUSTOM_ELEMENTS_SCHEMA]
 })
 export class BlogPageModule {}