+++ apps/client/src/app/pages/portfolio/portfolio-page.component.ts
@@ -3,16 +3,26 @@ import { TabConfiguration, User } from '@ghostfolio/common/interfaces';
 import { internalRoutes } from '@ghostfolio/common/routes/routes';
 
 import { ChangeDetectorRef, Component, OnDestroy, OnInit } from '@angular/core';
+import { MatTabsModule } from '@angular/material/tabs';
+import { RouterModule } from '@angular/router';
+import { addIcons } from 'ionicons';
+import {
+  analyticsOutline,
+  calculatorOutline,
+  pieChartOutline,
+  scanOutline,
+  swapVerticalOutline
+} from 'ionicons/icons';
 import { DeviceDetectorService } from 'ngx-device-detector';
 import { Subject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 
 @Component({
   host: { class: 'page has-tabs' },
+  imports: [MatTabsModule, RouterModule],
   selector: 'gf-portfolio-page',
   styleUrls: ['./portfolio-page.scss'],
-  templateUrl: './portfolio-page.html',
-  standalone: false
+  templateUrl: './portfolio-page.html'
 })
 export class PortfolioPageComponent implements OnDestroy, OnInit {
   public deviceType: string;
@@ -64,6 +74,14 @@ export class PortfolioPageComponent implements OnDestroy, OnInit {
           this.changeDetectorRef.markForCheck();
         }
       });
+
+    addIcons({
+      analyticsOutline,
+      calculatorOutline,
+      pieChartOutline,
+      scanOutline,
+      swapVerticalOutline
+    });
   }
 
   public ngOnInit() {