+++ libs/ui/src/lib/activity-type/activity-type.component.ts
@@ -8,11 +8,21 @@ import {
   Input,
   OnChanges
 } from '@angular/core';
+import { IonIcon } from '@ionic/angular/standalone';
 import { Type as ActivityType } from '@prisma/client';
+import { addIcons } from 'ionicons';
+import {
+  addCircleOutline,
+  arrowDownCircleOutline,
+  arrowUpCircleOutline,
+  cubeOutline,
+  flameOutline,
+  hammerOutline
+} from 'ionicons/icons';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [CommonModule],
+  imports: [CommonModule, IonIcon],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
   selector: 'gf-activity-type',
   styleUrls: ['./activity-type.component.scss'],
@@ -22,6 +32,17 @@ export class GfActivityTypeComponent implements OnChanges {
   @Input() activityType: ActivityType;
 
   public activityTypeLabel: string;
+
+  public constructor() {
+    addIcons({
+      addCircleOutline,
+      arrowDownCircleOutline,
+      arrowUpCircleOutline,
+      cubeOutline,
+      flameOutline,
+      hammerOutline
+    });
+  }
 
   public ngOnChanges() {
     this.activityTypeLabel = translate(this.activityType);