+++ libs/ui/src/lib/premium-indicator/premium-indicator.component.ts
@@ -8,10 +8,13 @@ import {
   Input
 } from '@angular/core';
 import { RouterModule } from '@angular/router';
+import { IonIcon } from '@ionic/angular/standalone';
+import { addIcons } from 'ionicons';
+import { diamondOutline } from 'ionicons/icons';
 
 @Component({
   changeDetection: ChangeDetectionStrategy.OnPush,
-  imports: [CommonModule, RouterModule],
+  imports: [CommonModule, IonIcon, RouterModule],
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
   selector: 'gf-premium-indicator',
   styleUrls: ['./premium-indicator.component.scss'],
@@ -21,4 +24,8 @@ export class GfPremiumIndicatorComponent {
   @Input() enableLink = true;
 
   public routerLinkPricing = publicRoutes.pricing.routerLink;
+
+  public constructor() {
+    addIcons({ diamondOutline });
+  }
 }